+++ src/main/scala/com/github/frankivo/CsvReader.scala
@@ -3,6 +3,8 @@ package com.github.frankivo
 import java.io.{BufferedReader, File}
 import java.time.LocalDate
 
+import com.github.frankivo.model.CovidRecord
+
 import scala.io.{BufferedSource, Source}
 import scala.util.Try
