+++ src/main/scala/com/github/frankivo/Graphs.scala
@@ -4,6 +4,7 @@ import java.nio.file.{Path, Paths}
 import java.time.LocalDate
 
 import akka.actor.Actor
+import com.github.frankivo.model.CovidRecord
 import com.github.frankivo.messages.{CreateMonthGraph, CreateWeeklyGraph}
 import org.jfree.chart.{ChartFactory, ChartUtils}
 import org.jfree.data.category.DefaultCategoryDataset
