+++ src/main/scala/com/github/frankivo/messages/CreateMonthGraph.scala
@@ -1,9 +1,10 @@
 package com.github.frankivo.messages
 
-import com.github.frankivo.CovidRecord
+import com.github.frankivo.model.CovidRecord
 
 /**
  * Make graph for month.
+ *
  * @param data Data of one month.
  */
 case class CreateMonthGraph(data: Seq[CovidRecord])
