+++ src/main/scala/com/github/frankivo/messages/RefreshData.scala
@@ -1,9 +1,10 @@
 package com.github.frankivo.messages
 
-import com.github.frankivo.CovidRecord
+import com.github.frankivo.model.CovidRecord
 
 /**
  * Pushes new data.
+ *
  * @param data The data.
  * @param containsUpdates True if any new data is present.
  */
