+++ src/main/scala/com/github/frankivo/model/Statistics.scala
@@ -1,4 +1,4 @@
-package com.github.frankivo
+package com.github.frankivo.model
 
 import java.time.LocalDate
